Fixed unit tests.

+++ geomesa-convert/geomesa-convert-common/src/main/scala/org/locationtech/geomesa/convert/Transformers.scala
@@ -391,7 +391,7 @@ class DateFunctionFactory extends TransformerFunctionFactory {
 
   val now                        = TransformerFn("now") { args => DateTime.now.toDate }
   val millisToDate               = TransformerFn("millisToDate") { args => new Date(args(0).asInstanceOf[Long]) }
-  val secsToDate                 = TransformerFn("secsToDate") { args => new Date(args(0).asInstanceOf[Int].toLong * 1000L) }
+  val secsToDate                 = TransformerFn("secsToDate") { args => new Date(args(0).asInstanceOf[Long] * 1000L) }
   val customFormatDateParser     = CustomFormatDateParser()
   val datetime                   = StandardDateParser("datetime", "dateTime")(ISODateTimeFormat.dateTime().withZoneUTC())
   val isodate                    = StandardDateParser("isodate", "basicDate")(ISODateTimeFormat.basicDate().withZoneUTC())
